Isolate activity visibility from request sessions
